move around and refactor type parameter
