+++ src/Generic.jl
@@ -32,6 +32,8 @@ include("generic/NCPoly.jl")
 
 include("generic/MPoly.jl")
 
+include("generic/PuiseuxMPoly.jl")
+
 include("generic/UniversalRing.jl")
 
 include("generic/SparsePoly.jl")
